requirements.txt
appnope==0.1.0
backcall==0.1.0
blessings==1.6.1
boto3==1.7.4
botocore==1.10.19
click==6.7
coverage==4.5.1
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.4
django-bootstrap4==0.0.6
django-configurations==2.0
django-filter==1.1.0
django-model-utils==3.1.1
django-nose==1.4.5
django-storages==1.6.6
django-unique-upload==0.2.1
djangorestframework==3.8.2
docutils==0.14
factory-boy==2.10.0
Faker==0.8.13
flake8==3.5.0
gunicorn==19.7.1
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
jmespath==0.9.3
livereload==2.5.2
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mkdocs==0.17.3
mock==2.0.0
newrelic==3.2.0.91
nose==1.3.7
nose-progressive==1.5.1
parso==0.2.0
pbr==4.0.2
pexpect==4.5.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
tornado==4.5.3
traitlets==4.3.2
wcwidth==0.1.7
